updating main with working app

+++ website/database/routines.sql
@@ -4,25 +4,28 @@ DELIMITER $$
 -- Functions
 --
 DROP FUNCTION IF EXISTS `add_move`$$
-CREATE DEFINER=`id15598586_root`@`%` FUNCTION `add_move` (`_playerCode` VARCHAR(20), `_gameCode` VARCHAR(10), `_position` INT, `_move` VARCHAR(10)) RETURNS VARCHAR(100) CHARSET utf8 COLLATE utf8_unicode_ci BEGIN
+CREATE FUNCTION `add_move` (`_playerCode` VARCHAR(20), `_gameCode` VARCHAR(10), `_position` INT, `_move` VARCHAR(10)) RETURNS VARCHAR(100) CHARSET utf8 COLLATE utf8_unicode_ci BEGIN
     DECLARE _affected_rows int default 0;
     DECLARE _field varchar(17);
     declare _temporary int default 0;
 	declare _dice_sum int default 0;
     declare _game_id int;
 	declare _field_tmp varchar(100);
+	declare player2 varchar(20);
+	declare _condition int default 0;
+	select 1 into _condition where _gameCode like "dices%";
     if _gameCode = "ttt" then
-        select field, COUNT(1) into _field, _affected_rows from current_matches where game_code='ttt' and (player_code_1=_playerCode or player_code_2=_playerCode);
+        select field, player_code_2, COUNT(1) into _field, player2, _affected_rows from current_matches where game_code='ttt' and player_code_1=_playerCode;
 		if _affected_rows > 0 then
             select insert(_field, _position, 1, _move) into _field_tmp;
 			set _field = _field_tmp;
-			update current_matches set field=_field where game_code=_gameCode and (player_code_1=_playerCode or player_code_2=_playerCode);
-            return "ok";
+			update current_matches set field=_field, player_code_1=player2, player_code_2=_playerCode where game_code=_gameCode and player_code_1=_playerCode;
+			return "ok";
         end if;
         return "request parameter not valid";
-    elseif _gameCode = "dices" then
+    elseif _condition != 0 then
         set _affected_rows = 0;
-        select id, count(1) into _game_id, _affected_rows from current_matches where game_code='dices' and (player_code_1=_playerCode or player_code_2=_playerCode);
+        select id, count(1) into _game_id, _affected_rows from current_matches where game_code=_gameCode and (player_code_1=_playerCode or player_code_2=_playerCode);
         if _affected_rows > 0 then
             set _affected_rows = 0;
             select count(1) into _affected_rows from last_move where game_id=_game_id and player_code=_playerCode;
@@ -55,19 +58,43 @@ CREATE DEFINER=`id15598586_root`@`%` FUNCTION `add_move` (`_playerCode` VARCHAR(
             return "cannot add more than one move";
         end if;
         return "request parameter not valid";
-    end if;
+    elseif _gameCode = "blow" then
+		set _affected_rows = 0;
+		select id, field, count(1) into _game_id, _field, _affected_rows from current_matches where game_code='blow' and (player_code_1=_playerCode or player_code_2=_playerCode);
+		if _affected_rows>0 then
+            select count(1) into _affected_rows from last_move where game_id=_game_id and player_code=_playerCode;
+			if _affected_rows <= 0 then
+				select `move`, count(1) into _field_tmp, _affected_rows from last_move where game_id=_game_id and player_code!=_playerCode; 
+				if _affected_rows > 0 then
+					delete from last_move where game_id=_game_id and player_code!=_playerCode; 
+					select convert(_field, signed) into _temporary;
+					if _temporary<=0 or _temporary>=200 then
+						return "match is already over";
+					end if;
+					select convert(_field_tmp, signed) into _dice_sum;
+					update current_matches set field=_temporary+((_position-_dice_sum)/10) where game_code=_gameCode and (player_code_1=_playerCode or player_code_2=_playerCode);
+					select field into _field from current_matches where game_code=_gameCode and (player_code_1=_playerCode or player_code_2=_playerCode);
+					return _field;
+				end if;
+				insert into last_move(game_id, player_code, `move`) values (_game_id, _playerCode, _position);
+				return "waiting for opponent move";
+			end if;
+			return "cannot add more than one move";
+		end if;
+		return "request parameter not valid";
+	end if;
     return "request parameter not valid";
 END$$
 
 DROP FUNCTION IF EXISTS `get_move`$$
-CREATE DEFINER=`id15598586_root`@`%` FUNCTION `get_move` (`_playerCode` VARCHAR(20), `_gameCode` VARCHAR(10)) RETURNS VARCHAR(100) CHARSET utf8 COLLATE utf8_unicode_ci BEGIN
+CREATE FUNCTION `get_move` (`_playerCode` VARCHAR(20), `_gameCode` VARCHAR(10)) RETURNS VARCHAR(100) CHARSET utf8 COLLATE utf8_unicode_ci BEGIN
     declare _id int;
     declare _affected_rows int;
     declare _field varchar(17);
     select id, count(1), field into _id, _affected_rows, _field from current_matches where game_code=_gameCode and (player_code_1=_playerCode or player_code_2=_playerCode);
 
     if _affected_rows>0 THEN
-        if _gameCode != "ttt" THEN
+        if _gameCode != "ttt" and _gameCode != "blow" THEN
             select move into _field from last_move where game_id=_id and player_code!=_playerCode;
             DELETE from last_move where game_id=_id and player_code!=_playerCode and move=_field;
         end if;
@@ -77,7 +104,7 @@ CREATE DEFINER=`id15598586_root`@`%` FUNCTION `get_move` (`_playerCode` VARCHAR(
 END$$
 
 DROP FUNCTION IF EXISTS `get_player`$$
-CREATE DEFINER=`id15598586_root`@`%` FUNCTION `get_player` (`_playerCode` VARCHAR(20), `_gameCode` VARCHAR(10)) RETURNS VARCHAR(100) CHARSET utf8 COLLATE utf8_unicode_ci BEGIN
+CREATE FUNCTION `get_player` (`_playerCode` VARCHAR(20), `_gameCode` VARCHAR(10)) RETURNS VARCHAR(100) CHARSET utf8 COLLATE utf8_unicode_ci BEGIN
 	DECLARE _id INT ;
 	DECLARE _affected_rows BOOLEAN ;
 	DECLARE _player_code varchar(20);
@@ -103,7 +130,7 @@ CREATE DEFINER=`id15598586_root`@`%` FUNCTION `get_player` (`_playerCode` VARCHA
 END$$
 
 DROP FUNCTION IF EXISTS `new_player`$$
-CREATE DEFINER=`id15598586_root`@`%` FUNCTION `new_player` (`_playerCode` VARCHAR(20), `_gameCode` VARCHAR(10)) RETURNS VARCHAR(100) CHARSET utf8 COLLATE utf8_unicode_ci BEGIN
+CREATE FUNCTION `new_player` (`_playerCode` VARCHAR(20), `_gameCode` VARCHAR(10)) RETURNS VARCHAR(100) CHARSET utf8 COLLATE utf8_unicode_ci BEGIN
     DECLARE _id INT DEFAULT 0; 
 	DECLARE _player_code VARCHAR(20) ; 
 	DECLARE _affected_rows BOOLEAN DEFAULT FALSE ;
@@ -126,7 +153,10 @@ CREATE DEFINER=`id15598586_root`@`%` FUNCTION `new_player` (`_playerCode` VARCHA
 		END IF ; 
 		IF _gameCode = "ttt" THEN 
 			SET _field = "-,-,-,-,-,-,-,-,-" ; 
-		ELSE 
+		
+		ELSEIF _gameCode = "blow" then
+			set _field = "100";
+		else 
 			SET _field = NULL ;
 		END IF ;
 		INSERT INTO current_matches(
